Update scripts for batch runs

+++ R/01-estimation.R
@@ -8,32 +8,37 @@ rm(list = ls())
 suppressMessages(library("EpiModelHIV"))
 suppressMessages(library("ARTnet"))
 
-epistats <- build_epistats(
-  geog.lvl = "city",
-  geog.cat = "Atlanta",
-  init.hiv.prev = c( 0.33, 0.137, 0.084),
-  race = TRUE
-)
-saveRDS(epistats, file = "data/input/epistats.rds")
-
-netparams <- build_netparams(epistats = epistats, smooth.main.dur = TRUE)
-netstats <- build_netstats(
-  epistats,
-  netparams,
-  expect.mort = 0.000478213,
-  network.size = 102000
-)
-saveRDS(netstats, file = "data/input/netstats.rds")
+ncores <- parallel::detectCores() - 1
 
 # 0. Initialize Network ---------------------------------------------------
 
-num <- netstats$demog$num
-nw <- network::network.initialize(num, directed = FALSE)
+if (!file.exists("data/input/netstats.rds")) {
+  epistats <- build_epistats(
+    geog.lvl = "city",
+    geog.cat = "Atlanta",
+    init.hiv.prev = c( 0.33, 0.137, 0.084),
+    race = TRUE
+  )
+  saveRDS(epistats, file = "data/input/epistats.rds")
 
-attr.names <- names(netstats$attr)
-attr.values <- netstats$attr
-nw <- network::set.vertex.attribute(nw, attr.names, attr.values)
-nw_main <- nw_casl <- nw_inst <- nw
+  netparams <- build_netparams(epistats = epistats, smooth.main.dur = TRUE)
+  netstats <- build_netstats(
+    epistats,
+    netparams,
+    expect.mort = 0.000478213,
+    network.size = 102000
+  )
+  saveRDS(netstats, file = "data/input/netstats.rds")
+
+  num <- netstats$demog$num
+  nw <- network::network_initialize(num, directed = FALSE)
+
+  attr.names <- names(netstats$attr)
+  attr.values <- netstats$attr
+  nw <- network::set_vertex_attribute(nw, attr.names, attr.values)
+  nw_main <- nw_casl <- nw_inst <- nw
+}
+
 
 # 1. Main Model -----------------------------------------------------------
 
@@ -69,12 +74,27 @@ fit_main <- netest(nw_main,
                    target.stats = netstats_main,
                    coef.diss = netstats$main$diss.byage,
                    set.control.ergm = control.ergm(MCMLE.maxit = 500,
-                                                   SAN.maxit = 3,
-                                                   SAN.nsteps.times = 3,
+                                                   SAN.maxit = 4,
+                                                   SAN.nsteps.times = 10,
                                                    MCMC.samplesize = 10000,
                                                    MCMC.interval = 5000,
-                                                   parallel = 30),
-                   verbose = TRUE)
+                                                   parallel = ncores),
+                   verbose = FALSE)
+
+# fit_main <- netest(nw_main,
+#                    formation = model_main,
+#                    target.stats = netstats_main,
+#                    coef.diss = netstats$main$diss.byage,
+#                    set.control.ergm = control.ergm(init.method = "MPLE",
+#                                            MCMLE.effectiveSize = NULL,
+#                                            MCMC.burnin = 1e6,
+#                                            MCMC.interval = 1e5,
+#                                            MCMC.samplesize = 10000,
+#                                            init.MPLE.samplesize = 2e8,
+#                                            MPLE.constraints.ignore = TRUE,
+#                                            parallel = 10,
+#                                            SAN.nsteps = 2e8),
+#                    verbose = TRUE)
 
 
 
@@ -112,8 +132,11 @@ fit_casl <- netest(nw_casl,
                    target.stats = netstats_casl,
                    coef.diss = netstats$casl$diss.byage,
                    set.control.ergm = control.ergm(MCMLE.maxit = 500,
-                                                   SAN.maxit = 3,
-                                                   SAN.nsteps.times = 3),
+                                                   SAN.maxit = 4,
+                                                   SAN.nsteps.times = 10,
+                                                   MCMC.samplesize = 10000,
+                                                   MCMC.interval = 5000,
+                                                   parallel = ncores),
                    verbose = FALSE)
 
 
@@ -149,8 +172,11 @@ fit_inst <- netest(nw_inst,
                    target.stats = netstats_inst,
                    coef.diss = dissolution_coefs(~offset(edges), 1),
                    set.control.ergm = control.ergm(MCMLE.maxit = 500,
-                                                   SAN.maxit = 3,
-                                                   SAN.nsteps.times = 3),
+                                                   SAN.maxit = 4,
+                                                   SAN.nsteps.times = 10,
+                                                   MCMC.samplesize = 10000,
+                                                   MCMC.interval = 5000,
+                                                   parallel = ncores),
                    verbose = FALSE)
